requirements.txt
astroid==2.2.5
coverage==4.5.3
Django==2.1.7
django-cors-headers==2.5.2
django-jet==1.0.8
django-webpack-loader==0.6.0
djangorestframework==3.9.2
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pylint==2.3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
whitenoise==4.1.2
wrapt==1.11.1
